requirements.txt
discord-py
aiosqlite
aiohttp